+++ packages/levit_di/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Added documentation for `LevitDisposable`.
+- Added example project.
+
 ## 0.0.1
 
 - Initial release.