+++ packages/levit_flutter/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Added documentation for `AsyncError`, `LScopedView`, `LState`, `LStatefulView`, and `LStatusBuilder`.
+- Added example application.
+
 ## 0.0.1
 
 - Initial release.