+++ packages/levit_reactive/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Added documentation for `AsyncError`, `LxAsyncComputed`, `LxHistoryMiddleware`, `LxMiddleware`, and `LxNotifier`.
+- Fixed lint issues in `core.dart`.
+- Added example project.
+
 ## 0.0.1
 
 - Initial release.